Update to 3.0z:
External compression updates. 3.0z prints a new warning when it
decompresses externally-compressed inputs, since I've received a
few reports of users confused by checksum failures. Remember, it
is not possible for xdelta3 to decompress and recompress a file
and ensure it has the same checksum. 3.0z improves error handling
for externally-compressed inputs with "trailing garbage" and also
includes a new flag to force the external compression command. Pass
'-F' to xdelta3 and it will pass '-f' to the external compression
command.