Update chat/telepathy-logger, math/gcalctool, print/evince, sysutils/brasero and sysutils/nautilus-sendto to use the new devel/glib2/schemas.mk logic to properly recompile the GSettings database during install/deinstall.
Bump PKGREVISION for these packages.

(This lets evince work in my machine; it would crash on startup because
it would not find its schema.  I suspect this also fixes all the other
applications mentioned in this change.  And the inclusion of schemas.mk
also prevents these packages from causing side-effects on the system if
they happen to call glib-compile-schemas during their build.)