Import droid-ttf-20100923 as fonts/droid-ttf
The Droid family of fonts was designed by Ascender's Steve Matteson
beginning in the fall of 2006. The goal was to provide optimal
quality and reading comfort on a mobile handset. The Droid fonts
were optimized for use in application menus, web browsers and for
other screen text.

Ascender Corporation worked closely with Google and the Open Handset
Alliance to develop these system fonts for Android, a free, open
source, and fully customizable mobile platform.